feat: move tool tests
