+++ bin/pytest-slow
@@ -3,4 +3,4 @@
 set -euo pipefail
 
 . .venv/bin/activate
-exec pytest "$@" -m 'not (tool or mujoco)' dimos
+exec pytest "$@" -m 'not mujoco' dimos
