+++ dimos/mapping/pointclouds/test_occupancy_speed.py
@@ -18,30 +18,10 @@ import time
 import pytest
 
 from dimos.mapping.pointclouds.occupancy import OCCUPANCY_ALGOS
-from dimos.mapping.voxels import VoxelGrid
-from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
 from dimos.utils.cli.plot import bar
-from dimos.utils.data import get_data, get_data_dir
-from dimos.utils.testing.replay import TimedSensorReplay
+from dimos.utils.data import get_data
 
 pytestmark = pytest.mark.self_hosted
-
-
-@pytest.mark.tool
-def test_build_map():
-    grid = VoxelGrid()
-
-    replay: TimedSensorReplay[PointCloud2] = TimedSensorReplay("go2_bigoffice/lidar")
-    for _ts, frame in replay.iterate_ts():
-        grid.add_frame(frame)
-
-    pickle_file = get_data_dir() / "unitree_go2_bigoffice_map.pickle"
-    global_pcd = grid.get_global_pointcloud2()
-
-    with open(pickle_file, "wb") as f:
-        pickle.dump(global_pcd, f)
-
-    grid.dispose()
 
 
 def test_costmap_calc():
