+++ dimos/mapping/test_voxels.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from collections.abc import Callable, Generator
-import time
 
 import numpy as np
 import pytest
@@ -68,45 +67,6 @@ def moment1(moment: MomentFactory) -> Go2MapperMoment:
 @pytest.fixture
 def moment2(moment: MomentFactory) -> Go2MapperMoment:
     return moment(85, False)
-
-
-@pytest.mark.tool
-def two_perspectives_loop(moment: MomentFactory) -> None:
-    while True:
-        moment(10, True)
-        time.sleep(1)
-        moment(85, True)
-        time.sleep(1)
-
-
-@pytest.mark.tool
-def test_carving(grid: VoxelGrid, moment1: Go2MapperMoment, moment2: Go2MapperMoment) -> None:
-    lidar_frame1 = moment1.lidar.value
-    assert lidar_frame1 is not None
-
-    lidar_frame2 = moment2.lidar.value
-    assert lidar_frame2 is not None
-
-    # Carving grid (default, carve_columns=True)
-    grid.add_frame(lidar_frame1)
-    grid.add_frame(lidar_frame2)
-    count_carving = grid.size()
-
-    voxel_size = grid._voxel_size
-    pts1 = np.asarray(lidar_frame1.pointcloud.points)
-    pts2 = np.asarray(lidar_frame2.pointcloud.points)
-    combined_vox = np.floor(np.vstack([pts1, pts2]) / voxel_size).astype(np.int64)
-    count_additive = np.unique(combined_vox, axis=0).shape[0]
-
-    print("\n=== Carving comparison ===")
-    print(f"Additive (no carving): {count_additive}")
-    print(f"With carving: {count_carving}")
-    print(f"Voxels carved: {count_additive - count_carving}")
-
-    # Carving should result in fewer voxels
-    assert count_carving < count_additive, (
-        f"Carving should remove some voxels. Additive: {count_additive}, Carving: {count_carving}"
-    )
 
 
 def test_ingest_a_few(grid: VoxelGrid) -> None:
