+++ dimos/memory2/test_voxel_map.py
@@ -14,27 +14,11 @@
 
 from __future__ import annotations
 
-import time
-from typing import TYPE_CHECKING
-
 import numpy as np
-import pytest
 
 from dimos.mapping.voxels import VoxelMapTransformer
-from dimos.memory2.store.sqlite import SqliteStore
 from dimos.memory2.type.observation import Observation
 from dimos.msgs.sensor_msgs.PointCloud2 import PointCloud2
-from dimos.utils.data import get_data
-
-if TYPE_CHECKING:
-    from collections.abc import Iterator
-
-
-@pytest.fixture(scope="module")
-def store() -> Iterator[SqliteStore]:
-    db = SqliteStore(path=get_data("go2_bigoffice.db"))
-    with db:
-        yield db
 
 
 def _make_obs(obs_id: int, points: np.ndarray, ts: float = 0.0) -> Observation[PointCloud2]:
@@ -103,33 +87,3 @@ def test_emit_every_n() -> None:
     assert results[0].tags["frame_count"] == 3
     assert results[1].tags["frame_count"] == 6
     assert results[2].tags["frame_count"] == 7
-
-
-# -- Integration tests against real replay data --
-
-
-@pytest.mark.tool
-def test_build_global_map(store: SqliteStore) -> None:
-    t_total = time.perf_counter()
-
-    lidar = store.stream("lidar", PointCloud2)
-    n_frames = lidar.count()
-
-    t0 = time.perf_counter()
-    result = lidar.transform(VoxelMapTransformer(voxel_size=0.05)).last()
-    t_transform = time.perf_counter() - t0
-
-    t_total = time.perf_counter() - t_total
-
-    global_map = result.data
-    frame_count = result.tags["frame_count"]
-
-    assert frame_count == n_frames
-    assert len(global_map) > 0
-
-    print(
-        lidar.summary(),
-        f"\n{frame_count} frames -> {len(global_map)} voxels"
-        f"\n  transform: {t_transform:.2f}s ({t_transform / frame_count * 1000:.1f}ms/frame)"
-        f"\n  total wall: {t_total:.2f}s",
-    )
