+++ dimos/msgs/sensor_msgs/tool_image.py
@@ -1,0 +1,31 @@
+# Copyright 2025-2026 Dimensional Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dimos.msgs.sensor_msgs.Image import Image
+from dimos.utils.data import get_data
+from dimos.utils.testing.legacy_pickle import LegacyPickleStore
+
+
+def test_sharpness_stream() -> None:
+    get_data("unitree_office_walk")  # Preload data for testing
+    video_store = LegacyPickleStore(
+        "unitree_office_walk/video", autocast=lambda x: Image.from_numpy(x).to_rgb()
+    )
+
+    cnt = 0
+    for image in video_store.iterate():
+        cnt = cnt + 1
+        print(image.sharpness)
+        if cnt > 30:
+            return
