+++ dimos/perception/detection/reid/tool_module.py
@@ -18,7 +18,6 @@ from dimos.perception.detection.reid.embedding_id_system import EmbeddingIDSyste
 from dimos.perception.detection.reid.module import ReidModule
 
 
-@pytest.mark.tool
 def test_reid_ingress(imageDetections2d) -> None:
     try:
         from dimos.models.embedding.treid import TorchReIDModel
