+++ docs/usage/transports/index.md
@@ -34,7 +34,7 @@ So: treat the API as uniform, but pick a backend whose semantics match the task.
 Quick view on performance of our pubsub backends:
 
 ```sh skip
-python -m pytest -svm tool -k "not bytes" dimos/protocol/pubsub/benchmark/test_benchmark.py
+python -m pytest -sv -k "not bytes" dimos/protocol/pubsub/benchmark/tool_benchmark.py
 ```
 
 ![Benchmark results](../assets/pubsub_benchmark.png)
@@ -461,7 +461,7 @@ See [`pubsub/test_spec.py`](/dimos/protocol/pubsub/test_spec.py) for the grid te
 Add your backend to benchmarks to compare in context:
 
 ```sh skip
-python -m pytest -svm tool -k "not bytes" dimos/protocol/pubsub/benchmark/test_benchmark.py
+python -m pytest -sv -k "not bytes" dimos/protocol/pubsub/benchmark/tool_benchmark.py
 ```
 
 ---
